+++ Controllers/AuthController.cs
@@ -4,16 +4,16 @@ using AuthServices.Models;
 using AuthServices.Services;
 using AuthServices.Services.Interfaces;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Identity.Data;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AuthServices.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class AuthController(IUserServices userServices, SignInManager<UsersEntity> signInManager) : ControllerBase
+public class AuthController(IUserService userService, SignInManager<UsersEntity> signInManager) : ControllerBase
 {
-    //private readonly UserManager<UsersEntity> _userManager;
-    private readonly IUserServices _userServices = userServices;
+    private readonly IUserService _userService = userService;
     private readonly SignInManager<UsersEntity> _signInManager = signInManager;
 
     [HttpPost("register")]
@@ -27,14 +27,14 @@ public class AuthController(IUserServices userServices, SignInManager<UsersEntit
             });
         }
 
-        var exists = await _userServices.UserExistsByEmailAsync(model.Email);
+        var exists = await _userService.UserExistsByEmailAsync(model.Email);
         if (exists)
             return Conflict("Email is already in use.");
 
         if (model.Password != model.ConfirmPassword)
             return BadRequest();
 
-        var created = await _userServices.CreateUser(model);
+        var created = await _userService.CreateUser(model);
 
         return created
              ? Ok("User registered successfully!")
@@ -51,7 +51,7 @@ public class AuthController(IUserServices userServices, SignInManager<UsersEntit
                 Errors = ModelState.Values.SelectMany(v => v.Errors).Select(e => e.ErrorMessage)
             });
         }
-        var user = await _userServices.GetUserByEmail(model.Email);
+        var user = await _userService.GetUserByEmail(model.Email);
         if (user == null)
             return Unauthorized("Invalid login attempt.");
 
@@ -68,13 +68,36 @@ public class AuthController(IUserServices userServices, SignInManager<UsersEntit
         return Ok("User signed out successfully!");
     }
 
+    [HttpPost("forgot-password")]
+    public async Task<IActionResult> RequestPasswordReset([FromBody] ForgotPasswordRequest request)
+    {
+        if (!ModelState.IsValid)
+            return BadRequest();
 
+        await _userService.RequestPasswordReset(request.Email);
+
+        //Always returns 200 Ok to prevent email enumeration
+        return Ok();
+    }
+
+    [HttpPost("reset-password")]
+    public async Task<IActionResult> ResetPassword([FromBody] ResetPasswordRequest request)
+    {
+        if (!ModelState.IsValid)
+            return BadRequest("ModelState not valid.");
+
+        var result = await _userService.ChangeForgottenPassword(request);
+
+        return result
+            ? Ok(new { success = true, message = "Password changed" })
+            : BadRequest(result);
+    }
 
     [HttpGet("GetUserById/{Id}")]
     public async Task<ActionResult<UserReturnData>> GetUserById(string Id)
     {
         if (Id == null) return BadRequest();
-        var user = await _userServices.GetUserById(Id);
+        var user = await _userService.GetUserById(Id);
         return Ok(user);
     }
 
@@ -82,7 +105,7 @@ public class AuthController(IUserServices userServices, SignInManager<UsersEntit
     public async Task<ActionResult<bool>> GetUserByEmail(string email)
     {
         if (string.IsNullOrEmpty(email)) return BadRequest();
-        var result = await _userServices.GetUserByEmail(email);
+        var result = await _userService.GetUserByEmail(email);
         return Ok(result);
     }
 
@@ -91,7 +114,7 @@ public class AuthController(IUserServices userServices, SignInManager<UsersEntit
     [HttpGet("GetAllUsers")]
     public async Task<ActionResult<IEnumerable<UserReturnData>>> GetAllUsers()
     {
-        var users = await _userServices.GetAllUsers();
+        var users = await _userService.GetAllUsers();
         return Ok(users);
     }
 
@@ -107,7 +130,7 @@ public class AuthController(IUserServices userServices, SignInManager<UsersEntit
                 Errors = ModelState.Values.SelectMany(v => v.Errors).Select(e => e.ErrorMessage)
             });
         }
-        var result = await _userServices.UpdateUser(user);
+        var result = await _userService.UpdateUser(user);
         return Ok(result);
         //if (result == true) return result("User updated succesfully");
     }
@@ -120,7 +143,7 @@ public class AuthController(IUserServices userServices, SignInManager<UsersEntit
         if (Id == null)
             return BadRequest();
 
-        return await _userServices.DeleteUser(Id);
+        return await _userService.DeleteUser(Id);
     }
